requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.18.0
azure-storage-queue==12.9.0
azure-storage-blob==12.19.0
chromadb==0.4.14
gremlinpython==3.6.4
langchain==0.1.0
langchain_community==0.0.10
langchain_core==0.1.22
langchain_openai==0.0.2
langgraph==0.0.24
psycopg2-binary==2.9.9
pydantic==1.10.2
Requests==2.29.0
aiohttp==3.8.3
ansi2html==1.9.1
moviepy==1.10.3